requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
magic-admin==2.2.1
redis==5.0.1
stripe==14.4.1
python-dotenv==1.0.0
pydantic[email]==2.5.0
python-dateutil==2.8.2
boto3==1.29.7